split into multiple crates to avoid cli dependency on openssl
