+++ crates/client/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "client"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+config = { path = "../config" }
+reqwest = { version = "0.11.27", features = [
+  "json",
+  "rustls-tls",
+], default-features = false }
+serde = { version = "1.0.197", features = ["derive"] }
+serde_json = "1.0.114"
