+++ crates/config/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "config"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+schemars = "0.8.16"
+serde = { version = "1.0.197", features = ["derive"] }
