+++ crates/config/src/lib.rs
@@ -1,0 +1,45 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct ServerConfigFile {
+    pub tables: Vec<TableConfig>,
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct ServerConfig {
+    /// the connection part of the config is not part of the config file
+    pub connection: ConnectionConfig,
+    pub tables: Vec<TableConfig>,
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct ConnectionConfig {
+    pub username: String,
+    pub password: String,
+    pub url: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct TableConfig {
+    pub name: String,
+    pub schema: String,
+    pub alias: String,
+    pub primary_key: Option<PrimaryKey>,
+    pub columns: Vec<ColumnConfig>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct PrimaryKey {
+    pub name: String,
+    pub columns: Vec<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
+pub struct ColumnConfig {
+    pub name: String,
+    pub alias: String,
+    pub data_type: String,
+}
+
+pub const CONFIG_FILE_NAME: &str = "configuration.json";
