+++ crates/ndc-clickhouse-cli/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ndc-clickhouse-cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap = { version = "4.5.3", features = ["derive", "env"] }
+client = { path = "../client" }
+config = { path = "../config" }
+serde = { version = "1.0.197", features = ["derive"] }
+serde_json = "1.0.114"
+tokio = { version = "1.36.0", features = ["macros", "rt-multi-thread", "fs"] }
