+++ crates/ndc-clickhouse-cli/src/database_introspection.sql
