+++ crates/ndc-clickhouse/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "ndc-clickhouse"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+async-trait = "0.1.78"
+client = { path = "../client" }
+config = { path = "../config" }
+indexmap = "2.1.0"
+ndc-sdk = { git = "https://github.com/hasura/ndc-hub.git", rev = "4cb500e", package = "ndc-sdk" }
+peg = "0.8.2"
+prometheus = "0.13.3"
+reqwest = { version = "0.11.27", features = [
+  "json",
+  "rustls-tls",
+], default-features = false }
+serde = { version = "1.0.197", features = ["derive"] }
+serde_json = "1.0.114"
+strum = { version = "0.26.2", features = ["derive"] }
+tokio = "1.36.0"
