+++ crates/ndc-clickhouse/src/connector.rs
@@ -1,24 +1,21 @@
-pub mod client;
-pub mod config;
 pub mod handler;
 pub mod state;
 
-use std::{env, fs::read_to_string, path::Path};
+use std::{env, path::Path};
+use tokio::fs;
 
 use async_trait::async_trait;
 use ndc_sdk::{
     connector::{
         Connector, ConnectorSetup, ExplainError, FetchMetricsError, HealthError,
-        InitializationError, MutationError, ParseError, QueryError, SchemaError,
+        InitializationError, LocatedError, MutationError, ParseError, QueryError, SchemaError,
     },
     json_response::JsonResponse,
     models,
 };
 
-use self::{
-    config::{read_server_config, ServerConfig},
-    state::ServerState,
-};
+use self::state::ServerState;
+use config::{ConnectionConfig, ServerConfig, ServerConfigFile, CONFIG_FILE_NAME};
 
 #[derive(Debug, Clone, Default)]
 pub struct ClickhouseConnector;
@@ -124,4 +121,51 @@ impl Connector for ClickhouseConnector {
             .map(JsonResponse::Value)
             .map_err(|err| QueryError::Other(err.to_string().into()))
     }
+}
+
+/// read server configuration from env var
+pub async fn read_server_config(
+    configuration_dir: impl AsRef<Path> + Send,
+) -> Result<ServerConfig, ParseError> {
+    let connection = get_connection_config()?;
+
+    let file_path = configuration_dir.as_ref().join(CONFIG_FILE_NAME);
+
+    let config_file = fs::read_to_string(&file_path)
+        .await
+        .map_err(|err| match err.kind() {
+            std::io::ErrorKind::NotFound => {
+                ParseError::CouldNotFindConfiguration(file_path.to_owned())
+            }
+            _ => ParseError::IoError(err),
+        })?;
+
+    let ServerConfigFile { tables } = serde_json::from_str::<ServerConfigFile>(&config_file)
+        .map_err(|err| {
+            ParseError::ParseError(LocatedError {
+                file_path,
+                line: err.line(),
+                column: err.column(),
+                message: err.to_string(),
+            })
+        })?;
+
+    Ok(ServerConfig { connection, tables })
+}
+
+fn get_connection_config() -> Result<ConnectionConfig, ParseError> {
+    // define what the new configuration will look like
+    // assemble config from env vars and reading files in config directory
+    let url = env::var("CLICKHOUSE_URL")
+        .map_err(|_err| ParseError::Other("CLICKHOUSE_URL env var must be set".into()))?;
+    let username = env::var("CLICKHOUSE_USERNAME")
+        .map_err(|_err| ParseError::Other("CLICKHOUSE_USERNAME env var must be set".into()))?;
+    let password = env::var("CLICKHOUSE_PASSWORD")
+        .map_err(|_err| ParseError::Other("CLICKHOUSE_PASSWORD env var must be set".into()))?;
+
+    Ok(ConnectionConfig {
+        url,
+        username,
+        password,
+    })
 }
