+++ crates/ndc-clickhouse/src/connector/handler/capabilities.rs
