+++ crates/ndc-clickhouse/src/schema/single_column_aggregate_function.rs
