+++ crates/ndc-clickhouse/src/sql.rs
