+++ crates/ndc-clickhouse/src/sql/query_builder/error.rs
